Update README.md
"This description is for the Persian (Farsi) version of the software. If you are using the non-Persian version, you may not need to install certain libraries."

In this revision, I've changed "Farsi" to "Persian" which is a more common term in English. I've also added "may" to indicate that it's not certain whether the libraries need to be installed in the non-Persian version, and I've added the word "certain" to clarify that it's only some libraries that may not need to be installed.